Refactor: Optimize async functions and clean up unused variables
- Converted several async functions to synchronous in commands:
  - `get_site_config`, `build_page_url`, `resolve_url`, `get_default_crawling_config`, `get_comprehensive_crawler_config`, `get_app_directories`, `is_first_run`, `cleanup_logs`, `get_log_directory_path` in `config_commands.rs`
  - `clear_crawling_errors`, `export_crawling_results` in `modern_crawling.rs`
  - `get_crawler_config`, `crawler_health_check` in `parsing_commands.rs`

- Cleaned up unused variables by prefixing with underscore in:
  - `actor_system.rs`
  - `system_analysis.rs`
  - `crawling_v4.rs`

- Removed unnecessary raw string literal hash symbols in various files.

- Eliminated trivial type casts in `database_connection.rs` and `dto.rs`.

- Added a script for automating the optimization process.

+++ src-tauri/src/application/dto.rs
@@ -152,8 +152,8 @@ impl From<MatterProduct> for MatterProductResponseDto {
     fn from(product: MatterProduct) -> Self {
         Self {
             url: product.url,
-            page_id: product.page_id.map(|id| id as u32),
-            index_in_page: product.index_in_page.map(|idx| idx as u32),
+            page_id: product.page_id.map(|id| id),
+            index_in_page: product.index_in_page.map(|idx| idx),
             id: product.id,
             manufacturer: product.manufacturer,
             model: product.model,
